refactor: 새롭게 만든 PageTemplate 사용
